--- test.txt
@@ -1,1 +1,0 @@
-test
